Enhanced #1
Added a `-a` command line or `create-people` option in the command line. When set, people not in the `people.json` file are still included in the Mardown files. Basically, it adds people missing from the config file. I think the `people.json` still needs at least one person in it.